winit: rework around `WinitEvent` vs `InputEvent<Special = WinitEvent>`
